Extend comm contexts to all comms
