+++ distributed/comm/tcp.py
@@ -12,7 +12,6 @@ except ImportError:
     ssl = None
 
 import dask
-import msgpack
 from tornado import netutil
 from tornado.iostream import StreamClosedError
 from tornado.tcpclient import TCPClient
@@ -26,9 +25,6 @@ from .registry import Backend, backends
 from .addressing import parse_host_port, unparse_host_port
 from .core import Comm, Connector, Listener, CommClosedError, FatalCommClosedError
 from .utils import to_frames, from_frames, get_tcp_server_address, ensure_concrete_host
-from ..protocol.compression import default_compression
-from ..protocol.utils import msgpack_opts
-from ..protocol import pickle
 
 
 logger = logging.getLogger(__name__)
@@ -352,10 +348,6 @@ class BaseTCPConnector(Connector, RequireEncryptionMixin):
                 ip, port, max_buffer_size=MAX_BUFFER_SIZE, **kwargs
             )
 
-            write_future = stream.write(msgpack.dumps(handshake_info()) + b"\n")
-            handshake = await stream.read_until(b"\n")
-            await write_future
-
             # Under certain circumstances tornado will have a closed connnection with an error and not raise
             # a StreamClosedError.
             #
@@ -367,19 +359,9 @@ class BaseTCPConnector(Connector, RequireEncryptionMixin):
             # The socket connect() call failed
             convert_stream_closed_error(self, e)
 
-        handshake = msgpack.loads(handshake.strip(), **msgpack_opts)
-
         local_address = self.prefix + get_stream_address(stream)
         comm = self.comm_class(
             stream, local_address, self.prefix + address, deserialize
-        )
-
-        comm.remote_info = handshake
-        comm.remote_info["address"] = self.prefix + address
-        comm.local_info = handshake_info()
-        comm.local_info["address"] = local_address
-        comm.handshake_options = handshake_configuration(
-            comm.local_info, comm.remote_info
         )
 
         return comm
@@ -468,20 +450,9 @@ class BaseTCPListener(Listener, RequireEncryptionMixin):
         local_address = self.prefix + get_stream_address(stream)
         comm = self.comm_class(stream, local_address, address, self.deserialize)
         comm.allow_offload = self.allow_offload
-        write_future = comm.stream.write(msgpack.dumps(handshake_info()) + b"\n")
-        handshake = await comm.stream.read_until(b"\n")
-        handshake = msgpack.loads(handshake.strip(), **msgpack_opts)
 
-        comm.remote_info = handshake
-        comm.remote_info["address"] = address
-        comm.local_info = handshake_info()
-        comm.local_info["address"] = local_address
+        await self.on_connection(comm)
 
-        comm.handshake_options = handshake_configuration(
-            comm.local_info, comm.remote_info
-        )
-
-        await write_future
         await self.comm_handler(comm)
 
     def get_host_port(self):
@@ -588,25 +559,6 @@ class TCPBackend(BaseTCPBackend):
 class TLSBackend(BaseTCPBackend):
     _connector_class = TLSConnector
     _listener_class = TLSListener
-
-
-def handshake_info():
-    return {
-        "compression": default_compression,
-        "python": tuple(sys.version_info)[:3],
-        "pickle-protocol": pickle.HIGHEST_PROTOCOL,
-    }
-
-
-def handshake_configuration(local, remote):
-    out = {"pickle-protocol": min(local["pickle-protocol"], remote["pickle-protocol"])}
-
-    if local["compression"] == remote["compression"]:
-        out["compression"] = local["compression"]
-    else:
-        out["compression"] = None
-
-    return out
 
 
 backends["tcp"] = TCPBackend()
