+++ distributed/comm/tests/test_comms.py
@@ -223,12 +223,11 @@ async def test_tcp_specific():
     assert host in ("localhost", "127.0.0.1", "::1")
     assert port > 0
 
-    connector = tcp.TCPConnector()
     l = []
 
     async def client_communicate(key, delay=0):
         addr = "%s:%d" % (host, port)
-        comm = await connector.connect(addr)
+        comm = await connect(listener.contact_address)
         assert comm.peer_address == "tcp://" + addr
         assert comm.extra_info == {}
         await comm.write({"op": "ping", "data": key})
@@ -270,12 +269,11 @@ async def test_tls_specific():
     assert host in ("localhost", "127.0.0.1", "::1")
     assert port > 0
 
-    connector = tcp.TLSConnector()
     l = []
 
     async def client_communicate(key, delay=0):
         addr = "%s:%d" % (host, port)
-        comm = await connector.connect(addr, ssl_context=client_ctx)
+        comm = await connect(listener.contact_address, ssl_context=client_ctx)
         assert comm.peer_address == "tls://" + addr
         check_tls_extra(comm.extra_info)
         await comm.write({"op": "ping", "data": key})
@@ -361,11 +359,10 @@ async def check_inproc_specific(run_client):
         == "inproc://" + listener_addr
     )
 
-    connector = inproc.InProcConnector(inproc.global_manager)
     l = []
 
     async def client_communicate(key, delay=0):
-        comm = await connector.connect(listener_addr)
+        comm = await connect(listener.contact_address)
         assert comm.peer_address == "inproc://" + listener_addr
         for i in range(N_MSGS):
             await comm.write({"op": "ping", "data": key})
@@ -682,13 +679,12 @@ async def check_comm_closed_implicit(addr, delay=None, listen_args={}, connect_a
         await comm.close()
 
     listener = await listen(addr, handle_comm, **listen_args)
-    contact_addr = listener.contact_address
 
-    comm = await connect(contact_addr, **connect_args)
+    comm = await connect(listener.contact_address, **connect_args)
     with pytest.raises(CommClosedError):
         await comm.write({})
 
-    comm = await connect(contact_addr, **connect_args)
+    comm = await connect(listener.contact_address, **connect_args)
     with pytest.raises(CommClosedError):
         await comm.read()
 
@@ -761,9 +757,8 @@ async def test_inproc_comm_closed_explicit_2():
             await comm.close()
 
     listener = await listen("inproc://", handle_comm)
-    contact_addr = listener.contact_address
 
-    comm = await connect(contact_addr)
+    comm = await connect(listener.contact_address)
     await comm.close()
     assert comm.closed()
     start = time()
@@ -777,7 +772,7 @@ async def test_inproc_comm_closed_explicit_2():
     with pytest.raises(CommClosedError):
         await comm.write("foo")
 
-    comm = await connect(contact_addr)
+    comm = await connect(listener.contact_address)
     await comm.write("foo")
     with pytest.raises(CommClosedError):
         await comm.read()
@@ -785,7 +780,7 @@ async def test_inproc_comm_closed_explicit_2():
         await comm.write("foo")
     assert comm.closed()
 
-    comm = await connect(contact_addr)
+    comm = await connect(listener.contact_address)
     await comm.write("foo")
 
     start = time()
